Make `Prop` generic
This makes it so props can be generic and return other types than string
:^)

+++ article-api/src/test/scala/no/ndla/articleapi/repository/ArticleRepositoryTest.scala
@@ -31,7 +31,7 @@ class ArticleRepositoryTest
   lazy val sampleArticle: Article = TestData.sampleArticleWithByNcSa
 
   def serverIsListening: Boolean = {
-    Try(new Socket(props.MetaServer.unsafeGet, props.MetaPort.unsafeGet.toInt)) match {
+    Try(new Socket(props.MetaServer.unsafeGet, props.MetaPort.unsafeGet)) match {
       case Success(c) =>
         c.close()
         true
