+++ concept-api/src/test/scala/no/ndla/conceptapi/repository/PublishedConceptRepositoryTest.scala
@@ -48,7 +48,7 @@ class PublishedConceptRepositoryTest extends IntegrationSuite(EnablePostgresCont
   }
 
   def serverIsListening: Boolean = {
-    Try(new Socket(props.MetaServer.unsafeGet, props.MetaPort.unsafeGet.toInt)) match {
+    Try(new Socket(props.MetaServer.unsafeGet, props.MetaPort.unsafeGet)) match {
       case Success(c) =>
         c.close()
         true
