+++ draft-api/src/test/scala/no/ndla/draftapi/repository/DraftRepositoryTest.scala
@@ -41,7 +41,7 @@ class DraftRepositoryTest extends IntegrationSuite(EnablePostgresContainer = tru
 
   def serverIsListening: Boolean = {
     val server = props.MetaServer.unsafeGet
-    val port   = props.MetaPort.unsafeGet.toInt
+    val port   = props.MetaPort.unsafeGet
     Try(new Socket(server, port)) match {
       case Success(c) =>
         c.close()
