+++ draft-api/src/test/scala/no/ndla/draftapi/repository/UserDataRepositoryTest.scala
@@ -38,7 +38,7 @@ class UserDataRepositoryTest extends IntegrationSuite(EnablePostgresContainer = 
 
   def serverIsListening: Boolean = {
     val server = props.MetaServer.unsafeGet
-    val port   = props.MetaPort.unsafeGet.toInt
+    val port   = props.MetaPort.unsafeGet
     Try(new Socket(server, port)) match {
       case Success(c) =>
         c.close()
