+++ myndla-api/src/test/scala/no/ndla/myndlaapi/repository/FolderRepositoryTest.scala
@@ -44,7 +44,7 @@ class FolderRepositoryTest
 
   def serverIsListening: Boolean = {
     val server = props.MetaServer.unsafeGet
-    val port   = props.MetaPort.unsafeGet.toInt
+    val port   = props.MetaPort.unsafeGet
     Try(new Socket(server, port)) match {
       case Success(c) =>
         c.close()
